backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose==3.3.0
pydantic==2.9.2
python-multipart==0.0.9   # <-- add this
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
